refactor(view): 自動スクロールのロジックをhooksとして切り出し
